+++ styles/theme/intensity.ts
@@ -1,66 +1,4 @@
 import chroma from "chroma-js"
-import { ThemeSyntax } from "@/src/themes/common/syntax"
-
-interface Author {
-    name: string
-    email: string
-    handle: string
-}
-
-type License = "MIT" | "Apache-2.0" | "GPL-3.0" | "Unlicense"
-
-enum Appearance {
-    Dark,
-    Light,
-}
-
-export type InputColor = string | string[]
-
-interface RequiredThemeProperties {
-    name: string
-    appearance: Appearance
-    author: string | Author
-    license: License
-    colors: {
-        neutral: InputColor
-        accent: InputColor
-        error: InputColor
-        info: InputColor
-        warning: InputColor
-        success: InputColor
-    }
-}
-
-interface OptionalThemeProperties {
-    url: string
-    syntax: ThemeSyntax
-}
-
-interface CalculatedThemeProperties {
-    intensity: {
-        min: number
-        max: number
-    }
-}
-
-export type ThemeConfig = RequiredThemeProperties &
-    Partial<OptionalThemeProperties>
-export type Theme = ThemeConfig & CalculatedThemeProperties
-
-const lightThemeConfig: ThemeConfig = {
-    appearance: "light",
-    calculatedIntensity: {
-        min: 5,
-        max: 88,
-    },
-}
-const darkThemeConfig: ThemeConfig = {
-    appearance: "dark",
-    calculatedIntensity: {
-        min: 1,
-        max: 92,
-    },
-}
 
 export function hexToIntensity(hex: string): number {
     const hsl = chroma(hex).hsl()
